use unint64_t instead of size_t for MACOS compatability

+++ src/transaction_builder.h
@@ -358,10 +358,10 @@ public:
     // Rust builders via ConvertRaw*(). The staging vectors above are cleared
     // after each convert call, so these counters preserve the committed counts
     // for use by IsValidSize().
-    size_t nCommittedSaplingSpends  = 0;
-    size_t nCommittedSaplingOutputs = 0;
-    size_t nCommittedOrchardSpends  = 0;
-    size_t nCommittedOrchardOutputs = 0;
+    uint64_t nCommittedSaplingSpends  = 0;
+    uint64_t nCommittedSaplingOutputs = 0;
+    uint64_t nCommittedOrchardSpends  = 0;
+    uint64_t nCommittedOrchardOutputs = 0;
 
 
     TransactionBuilder();
